remove turn_info class
All of its functions were moved to playmp_controller. Since they
were already so interconnected, this makes things a little easier.

Also it was always a bit confusing that this class was called
turn_info when all it did was to process network data.
